requirements.txt
# --- LangChain Core ---
langchain
langchain-core
langchain-openai
langchain-google-genai

# --- Vector Database Client ---
weaviate-client

# --- PDF & Markdown Processing ---
PyMuPDF # Para la librería fitz
markdown-it-py

# --- Evaluation Framework ---
ragas
datasets
tqdm

# --- Testing ---
pytest
pytest-asyncio
pytest-cov
pytest-mock

# --- Environment & Config ---
python-dotenv
pydantic
pydantic-settings
pyyaml
tenacity 

# --- Vector Database Client ---
supabase

# --- LangGraph Agent ---
langgraph
langchain-postgres
langgraph-checkpoint-postgres
psycopg[binary]
langgraph-cli[inmem]
langsmith
